+++ py2app_tests/basic_app/package2/sub/__init__.py
@@ -1,1 +1,1 @@
-""" package2.sub """
+foo = 42