+++ doc/changelog.rst
@@ -6,6 +6,9 @@ py2app 0.8
 
 py2app 0.8 is a feature release
 
+- Issue #39: It is now possible to have subpackages on
+  in the "packages" option of py2app.
+
 - Issue #37: Add recipe for pyEnchant
 
   ..note::